% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texto_wordpress.R
\name{cabecera_actividad}
\alias{cabecera_actividad}
\title{Genera la cabecera del post de wordpress de la actividad}
\usage{
cabecera_actividad(info)
}
\arguments{
\item{info}{información de la actividad (lista)}
}
\value{
cabecera del post
}
\description{
Genera la cabecera del post de wordpress de la actividad
}
