% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_concertadas.R
\name{completa_info_charlas_concertadas}
\alias{completa_info_charlas_concertadas}
\title{Completa la información de ubicación de las charlas concertadas.
Para las que vienen de solicitudes, valida que tengan referencia y nombre de centro y provincia concordantes.}
\usage{
completa_info_charlas_concertadas(dataset, solicitudes)
}
\arguments{
\item{dataset}{datos de charlas concertadas}

\item{solicitudes}{datos de solicitudes de charlas}
}
\value{
dataset de contactos con centros con columna "fallos_validacion" que indica si la petición es correcta
}
\description{
Completa la información de ubicación de las charlas concertadas.
Para las que vienen de solicitudes, valida que tengan referencia y nombre de centro y provincia concordantes.
}
