% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_contactos.R
\name{completa_info_contactos_centros}
\alias{completa_info_contactos_centros}
\title{Valida que las peticiones de contacto tengan referencia y nombre de centro concordantes.}
\usage{
completa_info_contactos_centros(dataset, solicitudes)
}
\arguments{
\item{dataset}{datos de contactos con centros}

\item{solicitudes}{datos de solicitudes de charlas}
}
\value{
dataset de contactos con centros con columna "validacion" que indica si la petición es correcta
}
\description{
Valida que las peticiones de contacto tengan referencia y nombre de centro concordantes.
}
