% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_aforo}
\alias{corrige_aforo}
\title{Corrección campo de aforo}
\usage{
corrige_aforo(aforos)
}
\arguments{
\item{aforos}{vector de aforos}
}
\value{
vector de aforos corregidos
}
\description{
Corrección campo de aforo
}
