% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_centro}
\alias{corrige_centro}
\title{Corrección campo de centro}
\usage{
corrige_centro(centros)
}
\arguments{
\item{centros}{vector de centros}
}
\value{
vector de centros corregidos
}
\description{
Corrección campo de centro
}
