% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_codigo_postal}
\alias{corrige_codigo_postal}
\title{Corrección campo de código postal}
\usage{
corrige_codigo_postal(cod_postales)
}
\arguments{
\item{cod_postales}{vector de códigos postales}
}
\value{
vector de código postales corregidos
}
\description{
Corrección campo de código postal
}
