% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_comentario}
\alias{corrige_comentario}
\title{Corrección campo de comentario}
\usage{
corrige_comentario(comentarios)
}
\arguments{
\item{comentarios}{vector de comentarios}
}
\value{
vector de comentarios corregidos
}
\description{
Corrección campo de comentario
}
