% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_comunidad_autonoma}
\alias{corrige_comunidad_autonoma}
\title{Corrección campo de comunidad autónoma}
\usage{
corrige_comunidad_autonoma(com_autonomas)
}
\arguments{
\item{com_autonomas}{vector de comunidades autónomas}
}
\value{
vector de comunidades autónomas corregidas
}
\description{
Corrección campo de comunidad autónoma
}
