% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_direccion}
\alias{corrige_direccion}
\title{Corrección campo de direccion}
\usage{
corrige_direccion(direcciones)
}
\arguments{
\item{direcciones}{vector de direcciones}
}
\value{
vector de direccions corregidas
}
\description{
Corrección campo de direccion
}
