% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_email}
\alias{corrige_email}
\title{Corrección campo de email}
\usage{
corrige_email(email)
}
\arguments{
\item{email}{vector de email}
}
\value{
vector de emails corregidos
}
\description{
Corrección campo de email
}
