% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_herramientas_online}
\alias{corrige_herramientas_online}
\title{Corrección campo de herramientas online}
\usage{
corrige_herramientas_online(herramientas_online)
}
\arguments{
\item{herramientas_online}{vector de herramientas online}
}
\value{
vector de enumeraciones de herramientas online corregidas
}
\description{
Corrección campo de herramientas online
}
