% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_imagen}
\alias{corrige_imagen}
\title{Corrección campo link de imagen}
\usage{
corrige_imagen(imagenes)
}
\arguments{
\item{imagenes}{vector de links a drive}
}
\value{
vector de imagenes corregidas
}
\description{
Corrección campo link de imagen
}
