% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_ingles}
\alias{corrige_ingles}
\title{Corrección campo de inglés}
\usage{
corrige_ingles(opciones_ingles)
}
\arguments{
\item{opciones_ingles}{vector de opciones de charla en inglés}
}
\value{
vector de opciones de charla en inglés corregidas
}
\description{
Corrección campo de inglés
}
