% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_localidad}
\alias{corrige_localidad}
\title{Corrección campo de localidad}
\usage{
corrige_localidad(localidades)
}
\arguments{
\item{localidades}{vector de localidades}
}
\value{
vector de localidades corregidas
}
\description{
Corrección campo de localidad
}
