% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_niveles}
\alias{corrige_niveles}
\title{Corrección campo de nivel}
\usage{
corrige_niveles(niveles)
}
\arguments{
\item{niveles}{vector de niveles}
}
\value{
vector de enumeraciones de niveles corregidas
}
\description{
Corrección campo de nivel
}
