% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_nombre}
\alias{corrige_nombre}
\title{Corrección campo de nombre}
\usage{
corrige_nombre(nombres)
}
\arguments{
\item{nombres}{vector de nombres}
}
\value{
vector de nombres corregidos
}
\description{
Corrección campo de nombre
}
