% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_numero}
\alias{corrige_numero}
\title{Corrección número entero}
\usage{
corrige_numero(numeros)
}
\arguments{
\item{tipos}{vector de números enteros}
}
\value{
vector de números corregido
}
\description{
Corrección número entero
}
