% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_provincia}
\alias{corrige_provincia}
\title{Corrección campo de provincia}
\usage{
corrige_provincia(provincias)
}
\arguments{
\item{provincias}{vector de provincias}
}
\value{
vector de provincias corregidas
}
\description{
Corrección campo de provincia
}
