% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_referencia}
\alias{corrige_referencia}
\title{Corrección campo de referencia}
\usage{
corrige_referencia(referencias)
}
\arguments{
\item{referencias}{vector de referencias}
}
\value{
vector de referencias corregidas
}
\description{
Corrección campo de referencia
}
