% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_si_no}
\alias{corrige_si_no}
\title{Corrección contestación Sí/No}
\usage{
corrige_si_no(si_no)
}
\arguments{
\item{tipos}{vector de contestaciones}
}
\value{
vector contestaciones corregidas
}
\description{
Corrección contestación Sí/No
}
