% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_telefono}
\alias{corrige_telefono}
\title{Corrección campo de teléfono}
\usage{
corrige_telefono(telefonos)
}
\arguments{
\item{telefonos}{vector de teléfonos}
}
\value{
vector de teléfonos corregidos
}
\description{
Corrección campo de teléfono
}
