% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_tipo_charla}
\alias{corrige_tipo_charla}
\title{Corrección tipo de charla concertada}
\usage{
corrige_tipo_charla(tipos)
}
\arguments{
\item{tipos}{vector de tipos de charla}
}
\value{
vector de tipos de charla corregidos
}
\description{
Corrección tipo de charla concertada
}
