% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_tipos}
\alias{corrige_tipos}
\title{Corrección campo de tipo}
\usage{
corrige_tipos(tipos)
}
\arguments{
\item{tipos}{vector de tipos}
}
\value{
vector de enumeraciones de tipos corregidas
}
\description{
Corrección campo de tipo
}
