% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_videollamada}
\alias{corrige_videollamada}
\title{Corrección campo de videollamada}
\usage{
corrige_videollamada(opciones_videollamada)
}
\arguments{
\item{opciones_videollamada}{vector de opciones de videollamada}
}
\value{
vector de opciones de videollamada corregidas
}
\description{
Corrección campo de videollamada
}
