% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_correccion.R
\name{corrige_web}
\alias{corrige_web}
\title{Corrección campo de web}
\usage{
corrige_web(webs)
}
\arguments{
\item{webs}{vector de webs}
}
\value{
vector de webs corregidas
}
\description{
Corrección campo de web
}
