% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tratamiento_duplicados.R
\name{elimina_duplicados}
\alias{elimina_duplicados}
\title{Elimina duplicados del dataset de solicitudes}
\usage{
elimina_duplicados(dataset)
}
\arguments{
\item{dataset}{dataset con los datos de las solicitudes en bruto}
}
\value{
dataset corregido
}
\description{
Elimina duplicados del dataset de solicitudes
}
