% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_actividades.R
\name{es_duplicado_actividad}
\alias{es_duplicado_actividad}
\title{Indica las actividades que están duplicadas,
manteniendo la última como la correcta}
\usage{
es_duplicado_actividad(dataset)
}
\arguments{
\item{dataset}{dataset con los datos de las actividades}
}
\value{
vector lógico indicando si cada petición es un duplicado o no
}
\description{
Indica las actividades que están duplicadas,
manteniendo la última como la correcta
}
