% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_contactos.R
\name{es_duplicado_contacto_centro}
\alias{es_duplicado_contacto_centro}
\title{Indica las peticiones de contacto con centro que son duplicados,
manteniendo la última como la correcta}
\usage{
es_duplicado_contacto_centro(dataset)
}
\arguments{
\item{dataset}{dataset con los datos de los contactos con centros}
}
\value{
vector lógico indicando si cada petición es un duplicado o no
}
\description{
Indica las peticiones de contacto con centro que son duplicados,
manteniendo la última como la correcta
}
