% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valida_referencia}
\alias{es_valida_referencia}
\title{Validación campo de referencia}
\usage{
es_valida_referencia(referencias)
}
\arguments{
\item{referencias}{vector de referencias}
}
\value{
vector de booleans que indica si cada formato de referencia es válido o no
}
\description{
Validación campo de referencia
}
