% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_aforo}
\alias{es_valido_aforo}
\title{Validación campo de aforo}
\usage{
es_valido_aforo(aforos)
}
\arguments{
\item{aforos}{vector de aforos}
}
\value{
vector de booleans que indica si cada aforo es válido o no
}
\description{
Validación campo de aforo
}
