% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_centro}
\alias{es_valido_centro}
\title{Validación campo de centro}
\usage{
es_valido_centro(centros)
}
\arguments{
\item{centros}{vector de centros}
}
\value{
vector de booleans que indica si cada centro es válido o no
}
\description{
Validación campo de centro
}
