% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_codigo_postal}
\alias{es_valido_codigo_postal}
\title{Validación campo de código postal}
\usage{
es_valido_codigo_postal(cod_postales)
}
\arguments{
\item{cod_postales}{vector de códigos postales}
}
\value{
vector de booleans que indica si cada código postal es válido o no
}
\description{
Validación campo de código postal
}
