% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_comentario}
\alias{es_valido_comentario}
\title{Validación campo de comentario}
\usage{
es_valido_comentario(comentarios)
}
\arguments{
\item{comentarios}{vector de comentarios}
}
\value{
vector de booleans que indica si cada comentario es válido o no
}
\description{
Validación campo de comentario
}
