% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_comunidad_autonoma}
\alias{es_valido_comunidad_autonoma}
\title{Validación campo de comunidad autónoma}
\usage{
es_valido_comunidad_autonoma(com_autonomas)
}
\arguments{
\item{com_autonomas}{vector de comunidades autónomas}
}
\value{
vector de booleans que indica si cada comunidad autónoma es válida o no
}
\description{
Validación campo de comunidad autónoma
}
