% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_direccion}
\alias{es_valido_direccion}
\title{Validación campo de direccion}
\usage{
es_valido_direccion(direcciones)
}
\arguments{
\item{direcciones}{vector de direcciones}
}
\value{
vector de booleans que indica si cada direccion es válida o no
}
\description{
Validación campo de direccion
}
