% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_email}
\alias{es_valido_email}
\title{Validación campo de email}
\usage{
es_valido_email(email)
}
\arguments{
\item{email}{vector de email}
}
\value{
vector de booleans que indica si cada email es válido o no
}
\description{
Validación campo de email
}
