% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_herramientas_online}
\alias{es_valido_herramientas_online}
\title{Validaci칩n campo de herramientas online}
\usage{
es_valido_herramientas_online(herramientas_online)
}
\arguments{
\item{herramientas_online}{vector de herramientas online}
}
\value{
vector de booleans que indica si cada enumeraci칩n de herramientas online es v치lida o no
}
\description{
Validaci칩n campo de herramientas online
}
