% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_imagen}
\alias{es_valido_imagen}
\title{Validación campo imagen}
\usage{
es_valido_imagen(imagenes)
}
\arguments{
\item{imagenes}{vector de aforos}
}
\value{
vector de booleans que indica si cada aforo es válido o no
}
\description{
Validación campo imagen
}
