% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_ingles}
\alias{es_valido_ingles}
\title{Validación campo de inglés}
\usage{
es_valido_ingles(opciones_ingles)
}
\arguments{
\item{opciones_ingles}{vector de opciones de charla en inglés}
}
\value{
vector de booleans que indica si cada opción de charla en inglés es válida o no
}
\description{
Validación campo de inglés
}
