% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_localidad}
\alias{es_valido_localidad}
\title{Validación campo de localidad}
\usage{
es_valido_localidad(localidades)
}
\arguments{
\item{localidades}{vector de localidades}
}
\value{
vector de booleans que indica si cada localidad es válida o no
}
\description{
Validación campo de localidad
}
