% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_niveles}
\alias{es_valido_niveles}
\title{Validaci칩n campo de nivel}
\usage{
es_valido_niveles(niveles)
}
\arguments{
\item{niveles}{vector de niveles}
}
\value{
vector de booleans que indica si cada enumeraci칩n de niveles es v치lida o no
}
\description{
Validaci칩n campo de nivel
}
