% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_nombre}
\alias{es_valido_nombre}
\title{Validación campo de nombre}
\usage{
es_valido_nombre(nombres)
}
\arguments{
\item{nombres}{vector de nombres}
}
\value{
vector de booleans que indica si cada nombre es válido o no
}
\description{
Validación campo de nombre
}
