% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_numero}
\alias{es_valido_numero}
\title{Validación número entero}
\usage{
es_valido_numero(numeros)
}
\arguments{
\item{tipos}{vector de números enteros}
}
\value{
vector de booleans que indica si cada número es válido o no
}
\description{
Validación número entero
}
