% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_provincia}
\alias{es_valido_provincia}
\title{Validación campo de provincia}
\usage{
es_valido_provincia(provincias)
}
\arguments{
\item{provincias}{vector de provincias}
}
\value{
vector de booleans que indica si cada provincia es válida o no
}
\description{
Validación campo de provincia
}
