% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_si_no}
\alias{es_valido_si_no}
\title{Validación contestación Sí/No}
\usage{
es_valido_si_no(si_no)
}
\arguments{
\item{tipos}{vector de contestaciones}
}
\value{
vector de booleans que indica si cada contestación es válida o no
}
\description{
Validación contestación Sí/No
}
