% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_telefono}
\alias{es_valido_telefono}
\title{Validación campo de teléfono}
\usage{
es_valido_telefono(telefonos)
}
\arguments{
\item{telefonos}{vector de teléfonos}
}
\value{
vector de booleans que indica si cada teléfono es válido o no
}
\description{
Validación campo de teléfono
}
