% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_tipo_charla}
\alias{es_valido_tipo_charla}
\title{Validación tipo de charla concertada}
\usage{
es_valido_tipo_charla(tipos)
}
\arguments{
\item{tipos}{vector de tipos de charla}
}
\value{
vector de booleans que indica si cada tipo de charla es válido o no
}
\description{
Validación tipo de charla concertada
}
