% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_tipos}
\alias{es_valido_tipos}
\title{Validaci칩n campo de tipo}
\usage{
es_valido_tipos(tipos)
}
\arguments{
\item{tipos}{vector de tipos}
}
\value{
vector de booleans que indica si cada enumeraci칩n de tipos es v치lida o no
}
\description{
Validaci칩n campo de tipo
}
