% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_videollamada}
\alias{es_valido_videollamada}
\title{Validaci칩n campo de videollamada}
\usage{
es_valido_videollamada(opciones_videollamada)
}
\arguments{
\item{opciones_videollamada}{vector de opciones de videollamada}
}
\value{
vector de booleans que indica si cada opci칩n de videollamada es v치lida o no
}
\description{
Validaci칩n campo de videollamada
}
