% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campos_validacion.R
\name{es_valido_web}
\alias{es_valido_web}
\title{Validación campo de web}
\usage{
es_valido_web(webs)
}
\arguments{
\item{webs}{vector de webs}
}
\value{
vector de booleans que indica si cada web es válida o no
}
\description{
Validación campo de web
}
