% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_datasets.R
\name{genera_id}
\alias{genera_id}
\title{Genera campo id en el dataset de acuerdo al timestamp de rellenado del formulario}
\usage{
genera_id(dataset, prefix = "")
}
\arguments{
\item{dataset}{tabla con los datos de formulario de googlesheets}

\item{prefix}{prefijo opcional que añadir a la ID}
}
\value{
dataset con un campo adicional para la id
}
\description{
Genera campo id en el dataset de acuerdo al timestamp de rellenado del formulario
}
