% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_actividades.R
\name{geolocaliza_actividades}
\alias{geolocaliza_actividades}
\title{Geolocaliza los centros del dataset de solicitudes de charlas}
\usage{
geolocaliza_actividades(dataset)
}
\arguments{
\item{dataset}{datos de actividades}
}
\value{
dataset de actividades geolocalizado
}
\description{
Geolocaliza los centros del dataset de solicitudes de charlas
}
