% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_datasets.R
\name{geolocaliza_centros}
\alias{geolocaliza_centros}
\title{Geolocaliza centros según sus datos postales
De momento lo único que hace es mirar las coordenadas del código postal
Si hay algún fallo de geolocalización informa en la columna de fallo}
\usage{
geolocaliza_centros(
  dataset,
  codpostales_file_id,
  col_codpostal,
  col_municipio,
  col_provincia,
  col_comautonoma,
  col_fallo
)
}
\arguments{
\item{dataset}{datos a geolocalizar}

\item{codpostales_file_id}{ID de Google Drive del fichero donde está el listado de códigos postales válidos}

\item{col_codpostal}{nombre de la columna con el código postal}

\item{col_municipio}{nombre de la columna con el municipio}

\item{col_provincia}{nombre de la columna con la provincia}

\item{col_comautonoma}{nombre de la columna con la comunidad autónoma}

\item{col_fallo}{nombre de la columna donde se registran los fallos}
}
\value{
dataset con nuevas columnas lon, lat con las coordenadas
}
\description{
Geolocaliza centros según sus datos postales
De momento lo único que hace es mirar las coordenadas del código postal
Si hay algún fallo de geolocalización informa en la columna de fallo
}
