% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_concertadas.R
\name{geolocaliza_charlas_concertadas}
\alias{geolocaliza_charlas_concertadas}
\title{Geolocaliza los centros del dataset de charlas concertadas}
\usage{
geolocaliza_charlas_concertadas(dataset)
}
\arguments{
\item{dataset}{datos de charlas concertadas}
}
\value{
dataset de charlas concertadas geolocalizado
}
\description{
Geolocaliza los centros del dataset de charlas concertadas
}
