% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_solicitudes.R
\name{geolocaliza_solicitudes_charlas}
\alias{geolocaliza_solicitudes_charlas}
\title{Geolocaliza los centros del dataset de solicitudes de charlas}
\usage{
geolocaliza_solicitudes_charlas(dataset)
}
\arguments{
\item{dataset}{datos de solicitudes de charlas}
}
\value{
dataset de solicitudes de charlas geolocalizado
}
\description{
Geolocaliza los centros del dataset de solicitudes de charlas
}
