% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_actividades.R
\name{get_actividades_limpio}
\alias{get_actividades_limpio}
\title{Get actividades (fichero limpio)}
\usage{
get_actividades_limpio(file_id, filename = NULL)
}
\arguments{
\item{file_id}{ID de Google Drive de la spreadsheet con los datos}

\item{filename}{Opcional. Nombre del fichero (incluyendo extensión ".csv") para guardar el dataset en disco.}
}
\value{
dataset limpio de actividades
}
\description{
Get actividades (fichero limpio)
}
