% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_actividades.R
\name{get_actividades_original}
\alias{get_actividades_original}
\title{Get actividades (fichero original)}
\usage{
get_actividades_original(file_id, filename = NULL)
}
\arguments{
\item{file_id}{ID de Google Drive de la spreadsheet con los datos}

\item{filename}{Opcional. Nombre del fichero (incluyendo extensión ".csv") para guardar el dataset en disco.}
}
\value{
dataset original de actividades
        con datos tal cual están en Google Drive
        y nombres de columnas adaptados
}
\description{
Get actividades (fichero original)
}
