% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_concertadas.R
\name{get_charlas_concertadas_limpio}
\alias{get_charlas_concertadas_limpio}
\title{Get charlas concertadas (fichero limpio)}
\usage{
get_charlas_concertadas_limpio(file_id, filename = NULL)
}
\arguments{
\item{file_id}{ID de Google Drive de la spreadsheet con los datos}

\item{filename}{Opcional. Nombre del fichero (incluyendo extensión ".csv") para guardar el dataset en disco.}
}
\value{
dataset limpio de charlas concertadas
}
\description{
Get charlas concertadas (fichero limpio)
}
