% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_concertadas.R
\name{get_charlas_concertadas_original}
\alias{get_charlas_concertadas_original}
\title{Get charlas concertadas (fichero original)}
\usage{
get_charlas_concertadas_original(file_id, filename = NULL)
}
\arguments{
\item{file_id}{ID de Google Drive de la spreadsheet con los datos}

\item{filename}{Opcional. Nombre del fichero (incluyendo extensión ".csv") para guardar el dataset en disco.}
}
\value{
dataset original de charlas concertadas
        con datos tal cual están en Google Drive
        y nombres de columnas adaptados
}
\description{
Get charlas concertadas (fichero original)
}
