% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_contactos.R
\name{get_contactos_centros_limpio}
\alias{get_contactos_centros_limpio}
\title{Get contactos con centros (fichero limpio)}
\usage{
get_contactos_centros_limpio(file_id, filename = NULL)
}
\arguments{
\item{file_id}{ID de Google Drive de la spreadsheet con los datos}

\item{filename}{Opcional. Nombre del fichero (incluyendo extensión ".csv") para guardar el dataset en disco.}
}
\value{
dataset limpio de contactos con centros
}
\description{
Get contactos con centros (fichero limpio)
}
