% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_contactos.R
\name{get_contactos_centros_original}
\alias{get_contactos_centros_original}
\title{Get contactos con centros (fichero original)}
\usage{
get_contactos_centros_original(file_id, filename = NULL)
}
\arguments{
\item{file_id}{ID de Google Drive de la spreadsheet con los datos}

\item{filename}{Opcional. Nombre del fichero (incluyendo extensión ".csv") para guardar el dataset en disco.}
}
\value{
dataset original de contactos con centros
        con datos tal cual están en Google Drive
        y nombres de columnas adaptados
}
\description{
Get contactos con centros (fichero original)
}
