% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googledrive.R
\name{get_drive_sheet}
\alias{get_drive_sheet}
\title{Get drive sheet}
\usage{
get_drive_sheet(file_id, filename = NULL)
}
\arguments{
\item{file_id}{ID de Google Drive de la spreadsheet con los datos}

\item{filename}{Opcional. Nombre del fichero (incluyendo extensión ".csv") para guardar el dataset en disco.}
}
\value{
dataset descargado tal cual está en Google Drive
}
\description{
Get drive sheet
}
