% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_solicitudes.R
\name{get_solicitudes_charlas_limpio}
\alias{get_solicitudes_charlas_limpio}
\title{Get solicitudes charlas (fichero limpio)}
\usage{
get_solicitudes_charlas_limpio(file_id, filename = NULL)
}
\arguments{
\item{file_id}{ID de Google Drive de la spreadsheet con los datos}

\item{filename}{Opcional. Nombre del fichero (incluyendo extensión ".csv") para guardar el dataset en disco.}
}
\value{
dataset limpio de solicitudes de charlas
}
\description{
Get solicitudes charlas (fichero limpio)
}
