% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_solicitudes.R
\name{get_solicitudes_charlas_original}
\alias{get_solicitudes_charlas_original}
\title{Get solicitudes charlas (fichero original)}
\usage{
get_solicitudes_charlas_original(file_id, filename = NULL)
}
\arguments{
\item{file_id}{ID de Google Drive de la spreadsheet con los datos}

\item{filename}{Opcional. Nombre del fichero (incluyendo extensión ".csv") para guardar el dataset en disco.}
}
\value{
dataset original de solicitudes de charlas
        con datos tal cual están en Google Drive
        y nombres de columnas adaptados
}
\description{
Get solicitudes charlas (fichero original)
}
