% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail.R
\name{gmail_acceso}
\alias{gmail_acceso}
\title{Autenticación de la cuenta de Gmail}
\usage{
gmail_acceso(credentials_path)
}
\arguments{
\item{credentials_path}{ruta al fichero JSON de credenciales}
}
\description{
Autenticación de la cuenta de Gmail
}
