% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail.R
\name{gmail_envia_email}
\alias{gmail_envia_email}
\title{Send email}
\usage{
gmail_envia_email(to, from, reply_to, subject, body)
}
\arguments{
\item{to}{email to}

\item{from}{email from}

\item{reply_to}{reply email to}

\item{subject}{email subject}

\item{body}{email body}
}
\description{
Send email
}
