% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texto_web.R
\name{html_link}
\alias{html_link}
\title{Genera link en HTML}
\usage{
html_link(texto, url, name = FALSE, new_tab = FALSE)
}
\arguments{
\item{texto}{texto a mostrar}

\item{url}{dirección a la que enlaza}

\item{name}{indica si crear una etiqueta 'name' o 'href'}

\item{new_tab}{indica si abrir el link en una nueva pestaña}
}
\value{
código HTML para el link
}
\description{
Genera link en HTML
}
