% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{leer_config}
\alias{leer_config}
\title{Leer fichero de configuración de la edición}
\usage{
leer_config(config_file)
}
\arguments{
\item{config_file}{ruta al fichero JSON de configuración}
}
\value{
objeto de configuración
}
\description{
Leer fichero de configuración de la edición
}
