% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_actividades.R
\name{limpia_actividades}
\alias{limpia_actividades}
\title{Limpia campos del dataset de actividades}
\usage{
limpia_actividades(dataset)
}
\arguments{
\item{dataset}{datos de actividades (de todas: centros, otras presenciales y otras no presenciales)}
}
\value{
dataset de actividades procesado
}
\description{
Limpia campos del dataset de actividades
}
