% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_concertadas.R
\name{limpia_charlas_concertadas}
\alias{limpia_charlas_concertadas}
\title{Limpia campos del dataset de charlas concertadas}
\usage{
limpia_charlas_concertadas(dataset)
}
\arguments{
\item{dataset}{datos de charlas concertadas}
}
\value{
dataset de charlas concertadas procesado
}
\description{
Limpia campos del dataset de charlas concertadas
}
