% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_contactos.R
\name{limpia_contactos_centros}
\alias{limpia_contactos_centros}
\title{Limpia campos del dataset de contactos con centros}
\usage{
limpia_contactos_centros(dataset)
}
\arguments{
\item{dataset}{datos de contactos con centros}
}
\value{
dataset de contactos con centros procesado
}
\description{
Limpia campos del dataset de contactos con centros
}
