% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_solicitudes.R
\name{limpia_solicitudes_charlas}
\alias{limpia_solicitudes_charlas}
\title{Limpia campos del dataset de solicitudes de charlas}
\usage{
limpia_solicitudes_charlas(dataset)
}
\arguments{
\item{dataset}{datos de solicitudes de charlas}
}
\value{
dataset de solicitudes de charlas procesado
}
\description{
Limpia campos del dataset de solicitudes de charlas
}
