% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texto_web.R
\name{link_referencia}
\alias{link_referencia}
\title{Genera link a referencia}
\usage{
link_referencia(id, texto, base_url)
}
\arguments{
\item{id}{número de referencia}

\item{texto}{texto a mostrar}

\item{base_url}{url de la página de referencias}
}
\value{
código HTML para el link
}
\description{
Genera link a referencia
}
