% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tratamiento_duplicados.R
\name{marca_duplicados}
\alias{marca_duplicados}
\title{Marca duplicados en el dataset de solicitudes}
\usage{
marca_duplicados(dataset)
}
\arguments{
\item{dataset}{tabla con los datos de solicitudes de googlesheets}
}
\value{
dataset marcado con DUP
}
\description{
Marca duplicados en el dataset de solicitudes
}
