% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_datasets.R
\name{original_vs_duplicados}
\alias{original_vs_duplicados}
\title{Indica el ID original en los registros duplicados
El original es el primer registro}
\usage{
original_vs_duplicados(dataset, campos_combrobacion)
}
\arguments{
\item{dataset}{dataset a comprobar}

\item{campos_combrobacion}{nombres de columnas a incluir en la comprobación}
}
\value{
vector con ID original si es duplicado, NA si no
}
\description{
Indica el ID original en los registros duplicados
El original es el primer registro
}
