% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texto_web.R
\name{popup_charla_concertada}
\alias{popup_charla_concertada}
\title{Genera popup a mostrar en el mapa de la web con informaci칩n de la charla concertada}
\usage{
popup_charla_concertada(info, base_url)
}
\arguments{
\item{info}{informaci칩n de la charla concertada (lista o dataframe)}

\item{base_url}{url de la p치gina de referencias}
}
\value{
popup con formato listo para mostrar en el mapa
}
\description{
Genera popup a mostrar en el mapa de la web con informaci칩n de la charla concertada
}
