% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_solicitudes.R
\name{referencia_duplicado_solicitud_charla}
\alias{referencia_duplicado_solicitud_charla}
\title{Indica las solicitudes de charla que son duplicados, devolviendo el ID de la correcta (la última)}
\usage{
referencia_duplicado_solicitud_charla(dataset, campos_combrobacion)
}
\arguments{
\item{dataset}{datos de solicitudes de charlas}
}
\value{
vector con ID de la solicitud correcta si es duplicado, NA si no
}
\description{
Indica las solicitudes de charla que son duplicados, devolviendo el ID de la correcta (la última)
}
