% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texto_web.R
\name{seccion_referencia}
\alias{seccion_referencia}
\title{Crea nueva sección para referencia en la página}
\usage{
seccion_referencia(id, texto)
}
\arguments{
\item{id}{número de referencia}

\item{texto}{texto a mostrar}
}
\value{
código HTML para el link
}
\description{
Crea nueva sección para referencia en la página
}
