% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texto_wordpress.R
\name{tags_actividad}
\alias{tags_actividad}
\title{Genera tags para etiquetar el post de wordpress de la actividad}
\usage{
tags_actividad(info, edicion)
}
\arguments{
\item{info}{información de la actividad (lista)}

\item{edicion}{año de la edición actual}
}
\value{
vector con las tags
}
\description{
Genera tags para etiquetar el post de wordpress de la actividad
}
