% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texto_wordpress.R
\name{texto_actividad}
\alias{texto_actividad}
\title{Genera el texto del post de wordpress de la actividad}
\usage{
texto_actividad(info)
}
\arguments{
\item{info}{información de la actividad (lista o dataframe)}
}
\value{
texto del post
}
\description{
Genera el texto del post de wordpress de la actividad
}
