% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texto_web.R
\name{texto_centro11f}
\alias{texto_centro11f}
\title{Genera texto a mostrar en la web con informaci칩n de la actividad de Centro11F}
\usage{
texto_centro11f(info)
}
\arguments{
\item{info}{informaci칩n de la actividad Centro11F (lista o dataframe)}

\item{base_url}{url de la p치gina de referencias}
}
\value{
texto con formato listo para imprimir
}
\description{
Genera texto a mostrar en la web con informaci칩n de la actividad de Centro11F
}
