% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texto_web.R
\name{texto_charla_concertada}
\alias{texto_charla_concertada}
\title{Genera texto a mostrar en la web con informaci칩n de la charla concertada}
\usage{
texto_charla_concertada(info)
}
\arguments{
\item{info}{informaci칩n de la charla concertada (lista o dataframe)}

\item{base_url}{url de la p치gina de referencias}
}
\value{
texto con formato listo para imprimir
}
\description{
Genera texto a mostrar en la web con informaci칩n de la charla concertada
}
