% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texto_email.R
\name{texto_email_contacto}
\alias{texto_email_contacto}
\title{Genera texto a mostrar en el email con informaci贸n de la persona que quiere contactar}
\usage{
texto_email_contacto(info)
}
\arguments{
\item{info}{informaci贸n de la petici贸n de contacto (lista o dataframe)}
}
\value{
texto con formato listo para imprimir
}
\description{
Genera texto a mostrar en el email con informaci贸n de la persona que quiere contactar
}
