% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texto_email.R
\name{texto_email_fallo}
\alias{texto_email_fallo}
\title{Genera texto a mostrar en el email de fallo con los datos introducidos del formulario}
\usage{
texto_email_fallo(info)
}
\arguments{
\item{info}{información de la petición de contacto (lista o dataframe)}
}
\value{
texto con formato listo para imprimir
}
\description{
Genera texto a mostrar en el email de fallo con los datos introducidos del formulario
}
