% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texto_web.R
\name{texto_solicitud_charla}
\alias{texto_solicitud_charla}
\title{Genera texto a mostrar en la web con informaci칩n de la solicitud}
\usage{
texto_solicitud_charla(info)
}
\arguments{
\item{info}{informaci칩n de la solicitud (lista o dataframe)}

\item{base_url}{url de la p치gina de referencias}
}
\value{
texto con formato listo para imprimir
}
\description{
Genera texto a mostrar en la web con informaci칩n de la solicitud
}
