% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_actividades.R
\name{upload_actividades_limpio}
\alias{upload_actividades_limpio}
\title{Upload actividades (fichero limpio)}
\usage{
upload_actividades_limpio(dataset, file_id, filename = NULL)
}
\arguments{
\item{dataset}{dataset limpio de contactos con centros}

\item{file_id}{ID de Google Drive de la spreadsheet con los datos}
}
\description{
Upload actividades (fichero limpio)
}
