% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_concertadas.R
\name{upload_charlas_concertadas_limpio}
\alias{upload_charlas_concertadas_limpio}
\title{Upload charlas concertadas (fichero limpio)}
\usage{
upload_charlas_concertadas_limpio(dataset, file_id)
}
\arguments{
\item{dataset}{dataset limpio de charlas concertadas}

\item{file_id}{ID de Google Drive de la spreadsheet con los datos}
}
\description{
Upload charlas concertadas (fichero limpio)
}
