% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_contactos.R
\name{upload_contactos_centros_limpio}
\alias{upload_contactos_centros_limpio}
\title{Upload contactos con centros (fichero limpio)}
\usage{
upload_contactos_centros_limpio(dataset, file_id)
}
\arguments{
\item{dataset}{dataset limpio de contactos con centros}

\item{file_id}{ID de Google Drive de la spreadsheet con los datos}
}
\description{
Upload contactos con centros (fichero limpio)
}
