% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googledrive.R
\name{upload_file_to_drive}
\alias{upload_file_to_drive}
\title{Upload/update local file to Drive folder}
\usage{
upload_file_to_drive(local_path, drive_folder_id)
}
\arguments{
\item{local_path}{ruta del fichero a subir}

\item{drive_folder_id}{ID de Google Drive de la carpeta donde subir el fichero}
}
\description{
Upload/update local file to Drive folder
}
