% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesamiento_solicitudes.R
\name{upload_solicitudes_charlas_limpio}
\alias{upload_solicitudes_charlas_limpio}
\title{Upload solicitudes charlas (fichero limpio)}
\usage{
upload_solicitudes_charlas_limpio(dataset, file_id)
}
\arguments{
\item{dataset}{dataset limpio de solicitudes de charlas}

\item{file_id}{ID de Google Drive de la spreadsheet con los datos}
}
\description{
Upload solicitudes charlas (fichero limpio)
}
